delete batch

+++ thirdai_python_package/neural_db/neural_db.py
@@ -569,11 +569,18 @@ class NeuralDB:
         on_success()
         return ids
 
-    def delete(self, source_id: str):
-        deleted_entities = self._savable_state.documents.delete(source_id)
+    def delete(self, source_ids: List[str]):
+        all_sources_exist = all(
+            source_id in self._savable_state.documents.registry
+            for source_id in source_ids
+        )
+        if not all_sources_exist:
+            raise ValueError("At least one source not found in neuraldb.")
+
+        deleted_entities = self._savable_state.documents.delete(source_ids)
         self._savable_state.model.delete_entities(deleted_entities)
         self._savable_state.logger.log(
-            session_id=self._user_id, action="delete", args={"source_id": source_id}
+            session_id=self._user_id, action="delete", args={"source_ids": source_ids}
         )
 
     def clear_sources(self) -> None:
